+++ components/Common/Modal/index.ts
@@ -1,3 +1,4 @@
 export { default as SmallModal } from './SmallModal';
 export { default as LargeModal } from './LargeModal';
 export { default as SquareModal } from './SquareModal';
+export { default as ModalWrapper } from './ModalWrapper';
