+++ components/Search/Modal/index.ts
@@ -1,1 +1,2 @@
 export { default as ApplyCompleteModal } from './ApplyCompleteModal';
+export { default as RecommendModal } from './RecommendModal';
