--- components/Search/Recommend/index.ts
@@ -1,1 +1,0 @@
-export { default as Recommend } from './Recommend';
